Moves toast to separate package
